sistemazioner struttura progetto

+++ store-security/backend/src/main/java/com/store/security/store_security/repository/ArticleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepository extends CrudRepository<ArticleEntity, Integer> {
+public interface ArticleRepository extends CrudRepository<ArticleEntity, Long> {
 }
